npdm: Silence -Wformat warnings
One case was specifying a size smaller than the actual type being
printed, while the other was caused due to C's integral promotion rules.
Operation on types smaller than int, will be promoted to an int. In this
case, the mask was unnecessary anyways, so it can just be removed.